+++ GramAddict/core/handle_sources.py
@@ -133,7 +133,7 @@ def handle_blogger(
 def handle_blogger_from_file(
     self,
     device,
-    current_filename,
+    parameter_passed,
     current_job,
     storage,
     on_interaction,
@@ -143,12 +143,26 @@ def handle_blogger_from_file(
     need_to_refresh = True
     on_following_list = False
     limit_reached = False
-    if path.isfile(current_filename):
-        with open(current_filename, "r", encoding="utf-8") as f:
+
+    filename = parameter_passed.split(" ")[0]
+    try:
+        amount_of_users = get_value(parameter_passed.split(" ")[1], None, 10)
+    except IndexError:
+        amount_of_users = 10
+        logger.warning(
+            f"You didn't passed how many users should be processed from the list! Default is {amount_of_users} users."
+        )
+    if path.isfile(filename):
+        with open(filename, "r", encoding="utf-8") as f:
             usernames = [line.replace(" ", "") for line in f if line != "\n"]
         len_usernames = len(usernames)
-        logger.info(f"In this file there are {len_usernames} entries.")
+        if len_usernames < amount_of_users:
+            amount_of_users = len_usernames
+        logger.info(
+            f"In {filename} there are {len_usernames} entries, {amount_of_users} users will be processed."
+        )
         not_found = []
+        processed_users = 0
         try:
             for line, username_raw in enumerate(usernames, start=1):
                 username = username_raw.strip()
@@ -164,12 +178,18 @@ def handle_blogger_from_file(
                         )
                         self.session_state.totalUnfollowed += 1
                         limit_reached = self.session_state.check_limit(
-                            self.args, limit_type=self.session_state.Limit.UNFOLLOWS
+                            limit_type=self.session_state.Limit.UNFOLLOWS
                         )
+                        processed_users += 1
                     else:
                         not_found.append(username_raw)
                     if limit_reached:
                         logger.info("Unfollows limit reached.")
+                        break
+                    if processed_users == amount_of_users:
+                        logger.info(
+                            f"{processed_users} users have been unfollowed, going to the next job."
+                        )
                         break
                 else:
                     if storage.is_user_in_blacklist(username):
@@ -216,26 +236,30 @@ def handle_blogger_from_file(
                     ):
                         return
                     device.back()
+                    processed_users += 1
+                    if processed_users == amount_of_users:
+                        logger.info(
+                            f"{processed_users} users have been interracted, going to the next job."
+                        )
+                        return
         finally:
             if not_found:
                 with open(
-                    f"{os.path.splitext(current_filename)[0]}_not_found.txt",
+                    f"{os.path.splitext(filename)[0]}_not_found.txt",
                     mode="a+",
                     encoding="utf-8",
                 ) as f:
                     f.writelines(not_found)
             if self.args.delete_interacted_users and len_usernames != 0:
-                with atomic_write(
-                    current_filename, overwrite=True, encoding="utf-8"
-                ) as f:
+                with atomic_write(filename, overwrite=True, encoding="utf-8") as f:
                     f.writelines(usernames[line:])
     else:
         logger.warning(
-            f"File {current_filename} not found. You have to specify the right relative path from this point: {os.getcwd()}"
+            f"File {filename} not found. You have to specify the right relative path from this point: {os.getcwd()}"
         )
         return
 
-    logger.info(f"Interact with users in {current_filename} completed.")
+    logger.info(f"Interact with users in {filename} completed.")
     device.back()
 
 
@@ -462,17 +486,31 @@ def handle_posts(
         return
 
     post_description = ""
+    likes_failed = 0
     nr_same_post = 0
     nr_same_posts_max = 3
     nr_consecutive_already_interacted = 0
+    already_liked_count = 0
+    already_liked_count_limit = 20
+    post_view_list = PostsViewList(device)
+    opened_post_view = OpenedPostView(device)
     while True:
-        flag, post_description, username, is_ad, is_hashtag = PostsViewList(
-            device
-        )._check_if_last_post(post_description, current_job)
-        has_likers, number_of_likers = PostsViewList(device)._find_likers_container()
-        already_liked, _ = OpenedPostView(device)._is_post_liked()
+        (
+            is_same_post,
+            post_description,
+            username,
+            is_ad,
+            is_hashtag,
+        ) = post_view_list._check_if_last_post(post_description, current_job)
+        has_likers, number_of_likers = post_view_list._find_likers_container()
+        already_liked, _ = opened_post_view._is_post_liked()
         if not (is_ad or is_hashtag):
-            if flag:
+            if already_liked_count == already_liked_count_limit:
+                logger.info(
+                    f"Limit of {already_liked_count_limit} already liked posts limit reached, finish."
+                )
+                break
+            if is_same_post:
                 nr_same_post += 1
                 logger.info(
                     f"Warning: {nr_same_post}/{nr_same_posts_max} repeated posts."
@@ -488,6 +526,7 @@ def handle_posts(
                 logger.info(
                     "Post already liked, SKIP.", extra={"color": f"{Fore.CYAN}"}
                 )
+                already_liked_count += 1
             elif random_choice(interact_percentage):
                 can_interact = False
                 if storage.is_user_in_blacklist(username):
@@ -524,34 +563,51 @@ def handle_posts(
                     logger.info(
                         f"Reached the limit of already interacted {skipped_posts_limit}. Goin to the next source/job!"
                     )
-                    return
+                    break
                 if can_interact and (likes_in_range or not has_likers):
                     logger.info(
                         f"@{username}: interact", extra={"color": f"{Fore.YELLOW}"}
                     )
                     if scraping_file is None:
-                        OpenedPostView(device).start_video()
-                        PostsViewList(device)._like_in_post_view(LikeMode.DOUBLE_CLICK)
-                        UniversalActions.detect_block(device)
-                        if not PostsViewList(device)._check_if_liked():
-                            PostsViewList(device)._like_in_post_view(
-                                LikeMode.SINGLE_CLICK
-                            )
+                        opened_post_view.start_video()
+                        if not session_state.check_limit(
+                            limit_type=session_state.Limit.LIKES, output=True
+                        ):
+                            post_view_list._like_in_post_view(LikeMode.DOUBLE_CLICK)
                             UniversalActions.detect_block(device)
-                        session_state.totalLikes += 1
-                        if current_job == "feed":
-                            count += 1
-                            logger.info(
-                                f"Interacted feed bloggers: {count}/{count_feed_limit}"
-                            )
-                            if count >= count_feed_limit:
-                                logger.info(
-                                    f"Interacted {count} bloggers in feed, finish."
-                                )
-                                TabBarView(device).navigateToProfile()
-                                return
+                            liked = post_view_list._check_if_liked()
+                            if not liked:
+                                post_view_list._like_in_post_view(LikeMode.SINGLE_CLICK)
+                                UniversalActions.detect_block(device)
+                                liked = post_view_list._check_if_liked()
+                            if liked:
+                                session_state.totalLikes += 1
+                                if current_job == "feed":
+                                    count += 1
+                                    logger.info(
+                                        f"Interacted feed bloggers: {count}/{count_feed_limit}"
+                                    )
+                                    likes_limit = self.session_state.check_limit(
+                                        limit_type=self.session_state.Limit.LIKES
+                                    )
+                                    success_limit = self.session_state.check_limit(
+                                        limit_type=self.session_state.Limit.SUCCESS
+                                    )
+                                    total_limit = self.session_state.check_limit(
+                                        limit_type=self.session_state.Limit.TOTAL
+                                    )
+                                    if likes_limit or success_limit or total_limit:
+                                        logger.info("Limit reached, finish.")
+                                        break
+                                    if count >= count_feed_limit:
+                                        logger.info(
+                                            f"Interacted {count} bloggers in feed, finish."
+                                        )
+                                        break
+                            else:
+                                likes_failed += 1
                     if current_job != "feed":
-                        opened, _, _ = PostsViewList(device)._post_owner(
+                        opened, _, _ = post_view_list._post_owner(
                             current_job, Owner.OPEN, username
                         )
                         if opened:
@@ -565,11 +621,14 @@ def handle_posts(
                                 current_job,
                                 on_interaction,
                             ):
-                                return
+                                break
                             device.back()
-
-        PostsViewList(device).swipe_to_fit_posts(SwipeTo.HALF_PHOTO)
-        PostsViewList(device).swipe_to_fit_posts(SwipeTo.NEXT_POST)
+        if likes_failed == 10:
+            logger.warning("You failed to do 10 likes! Soft-ban?!")
+            return
+        post_view_list.swipe_to_fit_posts(SwipeTo.HALF_PHOTO)
+        post_view_list.swipe_to_fit_posts(SwipeTo.NEXT_POST)
+    TabBarView(device).navigateToProfile()
 
 
 def handle_followers(