+++ GramAddict/core/utils.py
@@ -5,6 +5,7 @@ import re
 import shutil
 import subprocess
 import sys
+import time
 from datetime import datetime
 from math import nan
 from os import getcwd, rename, walk
@@ -12,6 +13,7 @@ from pathlib import Path
 from random import randint, shuffle, uniform
 from subprocess import PIPE
 from time import sleep
+from typing import Union
 from urllib.parse import urlparse
 
 import emoji
@@ -229,17 +231,34 @@ def kill_app(device, app_id):
     device.deviceV2.app_stop(app_id)
 
 
+def head_up_notifications(enabled: bool = False):
+    """
+    Enable or disable head-up-notifications
+    """
+    cmd = (
+        "adb"
+        + ("" if configs.device_id is None else " -s " + configs.device_id)
+        + f" shell settings put global heads_up_notifications_enabled {0 if not enabled else 1}"
+    )
+    return subprocess.run(cmd, stdout=PIPE, stderr=PIPE, shell=True, encoding="utf8")
+
+
 def open_instagram(device, screen_record, close_apps):
     nl = "\n"
     FastInputIME = "com.github.uiautomator/.FastInputIME"
     logger.info("Open Instagram app.")
-    cmd = (
-        "adb"
-        + ("" if configs.device_id is None else " -s " + configs.device_id)
-        + f" shell am start -n {app_id}/com.instagram.mainactivity.MainActivity"
-    )
-    cmd_res = subprocess.run(cmd, stdout=PIPE, stderr=PIPE, shell=True, encoding="utf8")
-    err = cmd_res.stderr.strip()
+
+    def call_ig():
+        cmd = (
+            "adb"
+            + ("" if configs.device_id is None else " -s " + configs.device_id)
+            + f" shell am start -n {app_id}/com.instagram.mainactivity.MainActivity"
+        )
+        return subprocess.run(
+            cmd, stdout=PIPE, stderr=PIPE, shell=True, encoding="utf8"
+        )
+
+    err = call_ig().stderr.strip()
     if "Error" in err:
         logger.error(err.replace(nl, ". "))
         return False
@@ -256,12 +275,14 @@ def open_instagram(device, screen_record, close_apps):
     max_tries = 3
     n = 0
     while device.deviceV2.info["currentPackageName"] != app_id:
-        if n > max_tries:
+        if n == max_tries:
             logger.critical("Unabled to open Instagram. Bot will stop.")
             return False
         n += 1
         logger.info(f"Waiting for Instagram to open... 😴 ({n}/{max_tries})")
-        check_if_crash_popup_is_there(device)
+        if check_if_crash_popup_is_there(device):
+            logger.info("Ig crashed, try to open it again...")
+        call_ig()
         random_sleep(3, 3, modulable=False)
 
     logger.info("Ready for botting!🤫", extra={"color": f"{Style.BRIGHT}{Fore.GREEN}"})
@@ -322,10 +343,75 @@ def close_instagram(device, screen_record):
             )
 
 
-def check_if_crash_popup_is_there(device):
-    obj = device.find(resourceId=ResourceID.AERR_RESTART)
+def check_if_crash_popup_is_there(device) -> bool:
+    obj = device.find(resourceId=ResourceID.CRASH_POPUP)
     if obj.exists():
         obj.click()
+        return True
+    return False
+
+
+def show_ending_conditions():
+    end_likes = configs.args.end_if_likes_limit_reached
+    end_follows = configs.args.end_if_follows_limit_reached
+    end_watches = configs.args.end_if_watches_limit_reached
+    end_comments = configs.args.end_if_comments_limit_reached
+    end_pm = configs.args.end_if_pm_limit_reached
+    logger.info(
+        "-" * 70,
+        extra={"color": f"{Style.BRIGHT}{Fore.YELLOW}"},
+    )
+    logger.info(
+        f"{'Session ending conditions:':<35} Value",
+        extra={"color": f"{Style.BRIGHT}{Fore.YELLOW}"},
+    )
+    logger.info(
+        "-" * 70,
+        extra={"color": f"{Style.BRIGHT}{Fore.YELLOW}"},
+    )
+    logger.info(
+        f"{'Likes:':<35} {end_likes}",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN if end_likes else Fore.RED}"},
+    )
+    logger.info(
+        f"{'Follows:':<35} {end_follows}",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN if end_follows else Fore.RED}"},
+    )
+    logger.info(
+        f"{'Watches:':<35} {end_watches}",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN if end_watches else Fore.RED}"},
+    )
+    logger.info(
+        f"{'Comments:':<35} {end_comments}",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN if end_comments else Fore.RED}"},
+    )
+    logger.info(
+        f"{'PM:':<35} {end_pm}",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN if end_pm else Fore.RED}"},
+    )
+    logger.info(
+        f"{'Total actions:':<35} True (not mutable)",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN}"},
+    )
+    logger.info(
+        f"{'Total successfull actions:':<35} True (not mutable)",
+        extra={"color": f"{Style.BRIGHT}{Fore.GREEN}"},
+    )
+    logger.info(
+        "For more info -> https://github.com/GramAddict/docs/blob/main/configuration.md#ending-session-conditions",
+        extra={"color": f"{Style.BRIGHT}{Fore.BLUE}"},
+    )
+    logger.info(
+        "-" * 70,
+        extra={"color": f"{Style.BRIGHT}{Fore.YELLOW}"},
+    )
+
+
+def countdown(seconds: int = 10, waiting_message: str = "") -> None:
+    while seconds:
+        print(waiting_message, f"{seconds:02d}", end="\r")
+        time.sleep(1)
+        seconds -= 1
 
 
 def pre_post_script(path: str, pre: bool = True):
@@ -424,6 +510,7 @@ def save_crash(device):
 def stop_bot(device, sessions, session_state, screen_record, was_sleeping=False):
     close_instagram(device, screen_record)
     kill_atx_agent(device)
+    head_up_notifications(enabled=True)
     logger.info(
         f"-------- FINISH: {datetime.now().strftime('%H:%M:%S')} --------",
         extra={"color": f"{Style.BRIGHT}{Fore.YELLOW}"},
@@ -433,10 +520,10 @@ def stop_bot(device, sessions, session_state, screen_record, was_sleeping=False)
         if not was_sleeping:
             sessions.persist(directory=session_state.my_username)
     ask_for_a_donation()
-    sys.exit(0)
+    sys.exit(2)
 
 
-def can_repeat(current_session, max_sessions):
+def can_repeat(current_session, max_sessions: int) -> bool:
     if max_sessions == -1:
         return True
     logger.info(
@@ -452,45 +539,35 @@ def can_repeat(current_session, max_sessions):
     return False
 
 
-def get_value(count, name, default, its_time=False):
-    def print_error():
+def get_value(
+    count: str, name: str, default: Union[int, float] = 0, its_time: bool = False
+) -> Union[int, float]:
+    def print_error() -> None:
         logger.error(
-            name.format(default)
-            + f'. Using default value instead of "{count}", because it must be '
+            f'Using default value instead of "{count}", because it must be '
             "either a number (e.g. 2) or a range (e.g. 2-4)."
         )
 
     parts = count.split("-")
-    if len(parts) <= 0:
-        value = default
-        print_error()
-    elif len(parts) == 1:
-        try:
+    try:
+        if len(parts) == 1:
             value = int(count)
-            if name is not None:
-                logger.info(name.format(value), extra={"color": Style.BRIGHT})
-        except ValueError:
-            value = default
-            print_error()
-    elif len(parts) == 2:
-        try:
+        elif len(parts) == 2:
             if not its_time:
                 value = randint(int(parts[0]), int(parts[1]))
             else:
                 value = round(uniform(int(parts[0]), int(parts[1])), 2)
-
-            if name is not None:
-                logger.info(name.format(value), extra={"color": Style.BRIGHT})
-        except ValueError:
-            value = default
-            print_error()
-    else:
+        else:
+            raise ValueError
+    except ValueError:
         value = default
         print_error()
+    if name is not None:
+        logger.info(name.format(value), extra={"color": Style.BRIGHT})
     return value
 
 
-def validate_url(x):
+def validate_url(x) -> bool:
     try:
         result = urlparse(x)
         return all([result.scheme, result.netloc, result.path])
@@ -499,7 +576,7 @@ def validate_url(x):
         return False
 
 
-def append_to_file(filename, username):
+def append_to_file(filename: str, username: str) -> None:
     try:
         if not filename.lower().endswith(".txt"):
             filename = filename + ".txt"