+++ GramAddict/plugins/interact_hashtag_likers.py
@@ -6,7 +6,6 @@ import emoji
 from colorama import Fore
 
 from GramAddict.core.decorators import run_safely
-from GramAddict.core.filter import Filter
 from GramAddict.core.handle_sources import handle_likers
 from GramAddict.core.interaction import (
     interact_with_user,
@@ -49,7 +48,7 @@ class InteractHashtagLikers(Plugin):
             },
         ]
 
-    def run(self, device, configs, storage, sessions, plugin):
+    def run(self, device, configs, storage, sessions, profile_filter, plugin):
         class State:
             def __init__(self):
                 pass
@@ -60,7 +59,6 @@ class InteractHashtagLikers(Plugin):
         self.sessions = sessions
         self.session_state = sessions[-1]
         self.args = configs.args
-        profile_filter = Filter(storage)
         self.current_mode = plugin
 
         # IMPORTANT: in each job we assume being on the top of the Profile tab already
@@ -74,10 +72,12 @@ class InteractHashtagLikers(Plugin):
         ]
         # Start
         for source in sample_sources(sources, self.args.truncate_sources):
-            limit_reached = self.session_state.check_limit(
-                self.args, limit_type=self.session_state.Limit.ALL
-            )
-
+            (
+                active_limits_reached,
+                _,
+                actions_limit_reached,
+            ) = self.session_state.check_limit(limit_type=self.session_state.Limit.ALL)
+            limit_reached = active_limits_reached or actions_limit_reached
             self.state = State()
             if source[0] != "#":
                 source = "#" + source
@@ -128,7 +128,7 @@ class InteractHashtagLikers(Plugin):
             if limit_reached:
                 logger.info("Ending session.")
                 self.session_state.check_limit(
-                    self.args, limit_type=self.session_state.Limit.ALL, output=True
+                    limit_type=self.session_state.Limit.ALL, output=True
                 )
                 break