+++ GramAddict/plugins/interact_place_likers.py
@@ -5,7 +5,6 @@ from random import seed
 from colorama import Style
 
 from GramAddict.core.decorators import run_safely
-from GramAddict.core.filter import Filter
 from GramAddict.core.handle_sources import handle_likers
 from GramAddict.core.interaction import (
     interact_with_user,
@@ -48,7 +47,7 @@ class InteractPlaceLikers(Plugin):
             },
         ]
 
-    def run(self, device, configs, storage, sessions, plugin):
+    def run(self, device, configs, storage, sessions, profile_filter, plugin):
         class State:
             def __init__(self):
                 pass
@@ -59,7 +58,6 @@ class InteractPlaceLikers(Plugin):
         self.sessions = sessions
         self.session_state = sessions[-1]
         self.args = configs.args
-        profile_filter = Filter(storage)
         self.current_mode = plugin
 
         # Handle sources
@@ -74,9 +72,12 @@ class InteractPlaceLikers(Plugin):
 
         # Start
         for source in sample_sources(sources, self.args.truncate_sources):
-            limit_reached = self.session_state.check_limit(
-                self.args, limit_type=self.session_state.Limit.ALL
-            )
+            (
+                active_limits_reached,
+                _,
+                actions_limit_reached,
+            ) = self.session_state.check_limit(limit_type=self.session_state.Limit.ALL)
+            limit_reached = active_limits_reached or actions_limit_reached
 
             self.state = State()
             logger.info(f"Handle {source}", extra={"color": f"{Style.BRIGHT}"})
@@ -123,7 +124,7 @@ class InteractPlaceLikers(Plugin):
             if limit_reached:
                 logger.info("Ending session.")
                 self.session_state.check_limit(
-                    self.args, limit_type=self.session_state.Limit.ALL, output=True
+                    limit_type=self.session_state.Limit.ALL, output=True
                 )
                 break